feat: named files & complete new party drag & drop
